safari-600.7-branch/LayoutTests/editing/inserting/5607069-2.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="div" contenteditable="true"><input type="text"><input type="text"></div>
<script>
Markup.description("This tests for a bug where InsertLineBreak would insert a '\\n' instead of a <br> if the caret was set just before an input field. There should be a br element between two input elements.");
div = document.getElementById("div");
window.getSelection().setPosition(div, 1);
document.execCommand("InsertLineBreak");
Markup.dump(div);
</script>
</body>
</html>
