third_party/blink/web_tests/editing/inserting/5607069-2-expected.txt

This tests for a bug where InsertLineBreak would insert a '\n' instead of a <br> if the caret was set just before an input field. There should be a br element between two input elements.
| <input>
|   type="text"
|   this.value=""
|   <shadow:root>
|     <div>
| <br>
| <input>
|   type="text"
|   this.value=""
|   <shadow:root>
|     <div>
